// src/models/aposta.ts
export interface Aposta {

  awayLogo: string;
  awayScore: number;
  awaySlug: string;
  awayTeam: string;
  formatedStartDate: string;
  homeLogo: string;
  homeScore: number;
  homeSlug: string;
  homeTeam: string;
  id: number;
  round: number;
  nickname: string;
  saved: true;
  slug: string;
  startTime: string;
  startTimeStamp: number;
  update: string;
  
}